refactor/extract class
