+++ lib/urbit/message.rb
@@ -57,16 +57,4 @@ module Urbit
       response
     end
   end
-
-  class CloseMessage < Message
-    def initialize(channel)
-      @action  = 'delete'
-      @channel = channel
-      @id      = 0
-    end
-
-    def to_h
-      {id: id, action: action}
-    end
-  end
 end
